requirements.txt
keras==2.2.4
tensorflow-gpu==2.0.0-beta1
Pillow==6.0.0
numpy
matplotlib
tqdm
scipy

